% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addSessionToPlan}
\alias{addSessionToPlan}
\title{Add a session to a plan}
\usage{
#add a run threshold to settings
mySettings <- addThreshold(sport = 'run', metric = 'pace', value = '4:30')

# create run session
runSession <- createSession(sport = 'run', metric = 'pace', minTargetZ = c(40, 20),
targetZ = c('4:50', '4:30'),
userSettings =  mySettings,
description = 'making a run test session')

# create a new plan
newPlan <- storeSettings(mySettings)

# add session to the plan
newPlan <- addSessionToPlan(weeklyPlan = newPlan, session = runSession,
day = 'thursday')
}
\arguments{
\item{weeklyPlan}{(weekPlan) An object with the weekly training plan. If NULL, a new one will be created}

\item{session}{(session) The new training session}

\item{day}{(character) The day of the week where to add the training session}
}
\description{
Add a training session to a new or existing weekly plan
}
