% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addThreshold.R
\name{addThreshold}
\alias{addThreshold}
\title{addThreshold}
\usage{
addThreshold(sport, metric, value, userSettings = NULL)
}
\arguments{
\item{sport}{Character. One of 'bike', 'run' or 'swim'}

\item{metric}{Character. One of 'power', pace', 'HR'}

\item{value}{Numeric (power and HR) or character (pace). For pace, use the format mm:ss}

\item{userSettings}{(userSettings) An object of class userSettings. Optional.
If given the new settings will be added to the existing ones, if NULL (default)
a new userSettings object will be created}
}
\description{
Function to set up a user threshold
}
\details{
Add threshold to user settings

The tool is adimentional, as in the calcualtions are not affected by the units used.
for instance min/km vs min/mile. However, the user should always use the same units, i.e.
if the threshold is in min/km, then all the sessions planned efforts must be entered in
min/km for the results to be correct.
}
\examples{
#set the bike FTP to 250W
mySettings <- addThreshold(sport = 'bike', metric = 'power', value = 250)

#add a run FTP of 300W
mySettings <- addThreshold(sport = 'bike', metric = 'power', value = 250,
userSettings = mySettings)
}
