% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cleanDay}
\alias{cleanDay}
\title{cleanDay}
\usage{
cleanDay(dayWeek)
}
\arguments{
\item{dayWeek}{An object of class dayWeek}
}
\value{
An object of calss dayWeek without the emoty session that could be generated as a
leftover of the create session
}
\description{
Clean a day from empty sessions
}
