% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cleanSettings}
\alias{cleanSettings}
\title{cleanSettings}
\usage{
cleanSettings(oldData, sport, metric)
}
\arguments{
\item{oldData}{data.frame of data to clean. From the slot settings in userSettings}

\item{sport}{character the sport to clean}

\item{metric}{character the metric to clean}
}
\description{
remove rows of already used sport/metric combination from userSettings data.frame
}
