% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convertToMinSec}
\alias{convertToMinSec}
\title{Convert a decimal to mm:ss string}
\usage{
convertToMinSec(thDec)
}
\arguments{
\item{thDec}{(numeric) a positve numeric}
}
\description{
Convert a positive integer to a string of mm:ss
}
