% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSession.R
\name{createSession}
\alias{createSession}
\title{Create a new training session}
\usage{
##set the swim threshold to 1:30
#myThreshold <- addThreshold(sport = 'swim', type = 'pace', value = '1:30')

##create a swim session
#newSession <- createSession(minTargetZ = c(20, 30), targetZ = c('1:20', '1:40'),
metric = 'pace', userSettings = myThreshold, type = 'swim',
description = 'making a test session')
}
\arguments{
\item{sport}{(character) One of the sports: bike, run, swim}

\item{metric}{(character) One of 'power', 'pace', 'HR'}

\item{minTargetZ}{(numeric) Number of a minutes at a specified intensity}

\item{targetZ}{(numeric, character) The session intensity for the minutes specified. Numeric for type = bike, character otherwise}

\item{userSettings}{(userSettings) an object of class userSettings}

\item{TSS}{(numeric) the known TSS for the session. If this is given (and is not NA (default))
it will be used to determine the session TSS, even if intervals are given.}

\item{description}{(character) Optional session description. If NULL, it will be set to the type}
}
\description{
Function to set up a user swim threshold
}
