% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesDef.R
\docType{class}
\name{dayWeek-class}
\alias{dayWeek-class}
\title{class for the day of the week session}
\description{
class for the day of the week session
}
\section{Slots}{

\describe{
\item{\code{sessions}}{list the list of session to do that day}
}}

