% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dayZones}
\alias{dayZones}
\title{dayZones}
\usage{
dayZones(object)
}
\arguments{
\item{object}{S4 object of class 'dayWeek'}
}
\value{
A dataframe of times in zones if it can be calcualted, NULL otherwise
}
\description{
function to get the total times in zones for a day
}
