% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{deleteSession}
\alias{deleteSession}
\title{Delete a session from a plan}
\usage{
##set the swim threshold to 1:30
#mySwimTh <- setSwimThreshold('1:30')

##create a swim session
#newSession <- createSession(c(20, 30), c('1:20', '1:40'), mySwimTh, 'making a test session')

##create the plan
#newPlan <- addSessionToPlan(weeklyPlan = NULL, session = newSession, day = 'monday', sessionNumber = 2)

##delete the session
#newPlan <- deleteSession(newPlan, 'monday', 2)
}
\arguments{
\item{weeklyPlan}{(weekPlan) An object with the weekly training plan. If NULL, a new one will be created}

\item{day}{(character) The day of the week where to add the training session}
}
\description{
Delete a training session from weekly plan
}
