% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDetails.R
\name{exportWeekDetails}
\alias{exportWeekDetails}
\title{exportWeekDetails}
\usage{
exportWeekDetails(weeklyPlan)
}
\arguments{
\item{weeklyPlan}{Object of class weeklyPlan}
}
\value{
Write a table to file
}
\description{
format a table with sesison details for export
}
