% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{formatZones}
\alias{formatZones}
\title{formatZones}
\usage{
formatZones(zones)
}
\arguments{
\item{zones}{A dataframe with 2 columns: zone (character) and time (numeric)}
}
\value{
a data.frame with xzones suummarized by name
}
\description{
Function to summarize the zones by zone name
}
