% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonesData.R
\name{getAllZones}
\alias{getAllZones}
\title{getAllZones}
\usage{
getAllZones()
}
\value{
A list of the zones
}
\description{
Contains the definition of training zones according to Matt Fitzgerald https://www.8020endurance.com/8020-zone-calculator/
}
