% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getDayTSS}
\alias{getDayTSS}
\title{Get day TSS}
\usage{
#add a run threshold to settings
mySettings <- addThreshold(sport = 'run', metric = 'pace', value = '4:30')

# create run session
runSession <- createSession(sport = 'run', metric = 'pace', minTargetZ = c(40, 20),
targetZ = c('4:50', '4:30'),
userSettings =  mySettings,
description = 'making a run test session')

# create a new plan
newPlan <- storeSettings(mySettings)

# add session to the plan
newPlan <- addSessionToPlan(weeklyPlan = newPlan, session = runSession,
day = 'thursday')

getDayTSS(slot(newPlan, 'thursday'))
}
\arguments{
\item{day}{(dayWeek) A dayWeek object}
}
\description{
Get day TSS
}
