% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getSessionDetails}
\alias{getSessionDetails}
\title{getSessionDetails}
\usage{
getSessionDetails(sport, metric, minutes, target, threshold, TSS, manualTSS)
}
\arguments{
\item{sport}{session sport}

\item{metric}{session metric}

\item{minutes}{minutes in target zone}

\item{target}{target zone}

\item{threshold}{threshold for the sport/metric}

\item{TSS}{session TSS}

\item{manualTSS}{if the TSS was user defined or not}
}
\description{
get the dataframe summarizing a session
}
