% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getSessionTSS}
\alias{getSessionTSS}
\title{Get single session TSS}
\usage{
##set the bike HR threshold to 170
myTh <- addThreshold(sport = 'bike', metric = 'HR', value = 170)

##create a bike session
newSession <- createSession(sport = 'bike', metric = 'HR', minTargetZc(20, 30),
targetZ = c(120, 150), userSettings = myTh, description = 'making a test session')

##get the session TSS
getSessionTSS(weeklyPlan)
}
\arguments{
\item{session}{(session) A session object}
}
\description{
Get single session TSS
}
