% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getSessionZones}
\alias{getSessionZones}
\title{getSessionZones}
\usage{
getSessionZones(sessionDetails, userSettings)
}
\arguments{
\item{sessionDetails}{(data.frame) The sesisonDetails data frame of a single session}

\item{userSettings}{(userSettings) The userSettings object}
}
\value{
A data.frame of times in zones. NULL if the time in zones cannot be calculated
}
\description{
Get the time in zones for a single session
}
