% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getThreshold}
\alias{getThreshold}
\title{getThreshold}
\usage{
getThreshold(sport, metric, userSettings)
}
\arguments{
\item{sport}{(character) The sport, i.e. 'bike'}

\item{metric}{(character) The metric, i.e. 'power'}

\item{userSettings}{(userSettings) The useSettings objects}
}
\description{
get the threshold valie from the user settings for the specified sport/metric
}
