% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getTSS}
\alias{getTSS}
\title{getTSS}
\usage{
getTSS(minTargetZ, targetZ, reference, metric)
}
\arguments{
\item{minTargetZ}{numeric number of minuts at a defined effort}

\item{targetZ}{numeric defined effort}

\item{reference}{effort threshold}

\item{metric}{the kind of threshold, i.e. power, HR or pace}
}
\description{
Function to calcualte TSS based on effort parameters and user threshold
}
