% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonesData.R
\name{getZones}
\alias{getZones}
\title{getZones}
\usage{
getZones(sport, metric)
}
\arguments{
\item{sport}{(character) The sport, i.e. bike}
}
\value{
A dataframe if the combination sport/metric exists, NULL otherwise
}
\description{
Get the specific datavrame of zones for a desired combination
of sport an metric. If the combination does not exist, the function returns NULL
}
