% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getZonesPercentages}
\alias{getZonesPercentages}
\title{getZonesPercentages}
\usage{
getZonesPercentages(object)
}
\arguments{
\item{object}{S4 object of class 'session', 'dayWeek' or 'weeklyPlan'}
}
\value{
A dataframe of times and percentage of times in zones if it
can be calcualted, NULL otherwise
}
\description{
function to get the total times in zones for a specified object
'session', 'dayWeek' or 'weeklyPlan'
}
