% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{moveSession}
\alias{moveSession}
\title{Reschedule a session within a plan}
\usage{
##set the swim threshold to 1:30
#mySwimTh <- setSwimThreshold('1:30')

##create a swim session
#newSession <- createSession(c(20, 30), c('1:20', '1:40'), mySwimTh, 'making a test session')

##create the plan
#newPlan <- addSessionToPlan(weeklyPlan = NULL, session = newSession, day = 'monday', sessionNumber = 2)

##move the session
#newPlan <- moveSession(weeklyPlan, 'monday', 2, 'tuesday', 1)
}
\arguments{
\item{fromDay}{(character) The day of the week where the training session  was scheduled}

\item{fromSessionNumber}{(numeric) The session number for fromDay, 1 or 2}

\item{toDay}{(character) The day of the week where to move the training session}

\item{toSessionNumber}{(numeric) The session number for that toDay, 1 or 2}

\item{replace}{(logic) If TRUE, replace the TO session with the FROM and the FROM with an epty session, if FALSE swaps sessions.}

\item{weekPlan}{(weekPlan) An object with the weekly training plan. If NULL, a new one will be created}
}
\description{
Add a training session to a new or existing weekly plan
}
