% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot,weeklyPlan,missing-method}
\alias{plot,weeklyPlan,missing-method}
\title{plot TSS distribution per day of the week}
\usage{
\S4method{plot}{weeklyPlan,missing}(x, y)
}
\arguments{
\item{x}{(weeklyPlan) An object of class dayWeek}

\item{y}{other paramenters}
}
\description{
plot TSS distribution per day of the week
}
