% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotTimeInZone}
\alias{plotTimeInZone}
\title{plotTimeInZone}
\usage{
plotTimeInZone(object)
}
\arguments{
\item{object}{S4 object of class 'session', 'dayWeek' or 'weeklyPlan'}
}
\value{
A ggplot
}
\description{
Make a barchart of the percentage of time spent in training zones
}
