% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesDef.R
\docType{class}
\name{singleSportSession-class}
\alias{singleSportSession-class}
\title{The individual training session}
\description{
The individual training session
}
\section{Slots}{

\describe{
\item{\code{sport}}{character of the sport}

\item{\code{metric}}{The metric used to calcualte TSS for this session}

\item{\code{TSS}}{numeric The session TSS}

\item{\code{manualTSS}}{logical TRUE if the TSS used was provided by the user}

\item{\code{sessionDetails}}{data.frame Details of each session interval}

\item{\code{zones}}{data.frame the time in training zones}
}}

