% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{storeSettings}
\alias{storeSettings}
\title{Store userSettings into a weeklyPlanner object}
\usage{
storeSettings(userSettings, weeklyPlan = NULL)
}
\arguments{
\item{userSettings}{(userSettings) An object of class userSettings}

\item{weeklyPlan}{(weeklyPlan) An object of class weeklyPlan}
}
\description{
Store userSettings into a weeklyPlanner object.
If the weeklyplan already has some sessions defined, the TSS will be recalculated
}
