% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{strToMinDec}
\alias{strToMinDec}
\title{strToMinDec}
\usage{
strToMinDec(inputStr)
}
\arguments{
\item{inputStr}{a vector of string or numeric}
}
\description{
Convert input to decimal. The input is
either already numeric or a string. If a string, the
format is either nn.nn or mm:ss
}
