% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSummary.R
\name{summary,dayWeek-method}
\alias{summary,dayWeek-method}
\title{define a summary method for dayWeek}
\usage{
\S4method{summary}{dayWeek}(object)
}
\arguments{
\item{object}{(dayWeek) An object of class dayWeek}
}
\description{
define a summary method for dayWeek
}
