% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSummary.R
\name{summary,session-method}
\alias{summary,session-method}
\title{define a summary method for session}
\usage{
\S4method{summary}{session}(object)
}
\arguments{
\item{object}{(session) An object of class session}
}
\description{
define a summary method for session
}
