% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSummary.R
\name{summary,singleSportSession-method}
\alias{summary,singleSportSession-method}
\title{define a summary method for session}
\usage{
\S4method{summary}{singleSportSession}(object)
}
\arguments{
\item{object}{(singleSportSession) An object of class session}
}
\description{
define a summary method for session
}
