% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSummary.R
\name{summary,userSettings-method}
\alias{summary,userSettings-method}
\title{define a summary method for usersettings}
\usage{
\S4method{summary}{userSettings}(object)
}
\arguments{
\item{object}{(userSettings) An object of class userSettings}
}
\description{
define a summary method for usersettings
}
