% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSummary.R
\name{summary,weeklyPlan-method}
\alias{summary,weeklyPlan-method}
\title{define a summary method for weeklyPlan}
\usage{
\S4method{summary}{weeklyPlan}(object)
}
\arguments{
\item{object}{(weeklyPlan) An object of class weeklyPlan}
}
\description{
define a summary method for weeklyPlan
}
