% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tableOfTss}
\alias{tableOfTss}
\title{Create a table of TSS for the week}
\usage{
tableOfTss(weeklyPlan)
}
\arguments{
\item{weeklyPlan}{(weeklyPlan) An object of class weeklyPlan}
}
\description{
Create a table of TSS for the week
}
