% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesDef.R
\docType{class}
\name{userSettings-class}
\alias{userSettings-class}
\title{userSettings for thresholds used to calculate TSS}
\description{
userSettings for thresholds used to calculate TSS
}
\section{Slots}{

\describe{
\item{\code{settings}}{data.frame of settings}
}}

