% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateTimes}
\alias{validateTimes}
\title{validateTimes}
\usage{
validateTimes(metric, values)
}
\arguments{
\item{metric}{session metric}

\item{values}{the valus to convert}
}
\description{
return the values in decima format
}
