% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{viewSessionDetails}
\alias{viewSessionDetails}
\title{viewSessionDetails}
\usage{
viewSessionDetails(session)
}
\arguments{
\item{session}{An object of calss session}
}
\description{
view the slot sessionDetails in a session object
}
