% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSession.R
\name{createSession}
\alias{createSession}
\title{createSession}
\usage{
createSession(
  sports,
  metrics = list(NA),
  minTargetZs = list(NA),
  targetZs = list(NA),
  TSS = list(NA),
  userSettings = NULL,
  description = ""
)
}
\arguments{
\item{sports}{(list) One of the sports: bike, run, swim}

\item{metrics}{(list) One of 'power', 'pace', 'HR'}

\item{minTargetZs}{(list) Number of a minutes at a specified intensity for each sport/session}

\item{targetZs}{(list) The session intensity for the minutes specified. Numeric for type = bike, character otherwise}

\item{TSS}{(list) the known TSS for the session. If this is given (and is not NA (default))
it will be used to set the session TSS, even if intervals are given.}

\item{userSettings}{(userSettings) an object of class userSettings}

\item{description}{(character) Optional session description. If NULL, it will be set to the type}
}
\description{
Function to set up a user swim threshold
}
\details{
Create a new training session
}
\examples{
##set the run threshold to 4:30
myThreshold <- addThreshold(sport = 'run', metric = 'pace', value = '4:30')
##set the bike threshold to 250
myThreshold <- addThreshold(sport = 'bike', metric = 'power', value = 250, userSettings = myThreshold)

##create a brick session
newSession <- createSession(sports=list('run', 'bike'),
minTargetZ = list(c(20, 30), 60), targetZ = list(c('1:20', '1:40'), 250),
metric = list('pace', 'power'), userSettings = myThreshold, TSS = list(NA, NA),
description = 'making a test session')
}
