% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesDef.R
\docType{class}
\name{session-class}
\alias{session-class}
\title{The individual training session}
\description{
The individual training session
}
\section{Slots}{

\describe{
\item{\code{sessions}}{list of sessions in the brick session}

\item{\code{description}}{A character description for the session}
}}

